Revert "[Refactor] Do not rely on implicit conversions in {fmt} calls  (#72022)"
This reverts commit bbef286d9d80969f3c9068cd98a04ba62090ed3c.

+++ be/src/base/compression/stream_decompressor.cpp
@@ -1150,7 +1150,7 @@ StatusOr<std::unique_ptr<StreamDecompressor>> StreamDecompressor::create_decompr
         decompressor = std::make_unique<LzoStreamDecompressor>();
         break;
     default:
-        return Status::InternalError(fmt::format("Unknown compress type: {}", static_cast<int>(type)));
+        return Status::InternalError(fmt::format("Unknown compress type: {}", type));
     }
 
     RETURN_IF_ERROR(decompressor->init());
